backend/requirements.txt
numpy
pandas
matplotlib
scikit-learn
tensorflow
ta
python-dotenv
requests
alpaca-trade-api
pydantic[email]
slowapi
pytest
pytest-asyncio
httpx
vaderSentiment